fix two bugs related to gerber file generation.  The first bug is triggered when the first aperture used in a file is the same as the last aperture used in the previous file.  In this case the aperture selection code is missing from the output file.
The second bug is when the first point drawn has its Y coordinate (in PCB
coordinates) equal to zero.  The output will be at gerber Y coordinate
zero which is on the opposite side of the board.

Thanks to Gabriel Paubert (paubert at iram dot es) for noting these bugs
and supplying a patch.