Add code to the parser which looks for a line like
FileVersion[20061101]

at the top of the .pcb file.  If this line is seen and indicates a version
which is too new for this version of pcb then issue a complaint and explain
why.  This should help out when the file format changes and someone tries to
load a new design with an older version of pcb.  Of course the real benefit
won't happen until versions of pcb prior to this check all go away...

For now, the actual output file from pcb hasn't changed.  There is a two line
change to include the FileVersion[] line that can be included the next time
the file format is changed.  However, there is no reason to not have the
program start watching for it.