Apply filters to load filechooser dialogs. [1988982] [2686963]
Applies filters to the filechooser dialogs when loading layouts,
layouts (to buffer), elements (to buffer) and netlists.

Default behaviour is to not filter in the filechooser dialog.
Choosing a predefined filefilter in the filechooser dialog filters
on registered mime types, lowercase and uppercase file extensions.
Predefined filters are selected upon the action chosen in the "File"
pulldown menu.