Some cleanup of configure script
Added top level check for pkg-config with PKG_PROG_PKG_CONFIG and removed
AC_PATH_PROG call for pkg-config from inside a block conditional on
enable_dbus.

Added a PKG_CHECK_MODULE for cairo if the toporouter output is enabled.

Added a check for HID dependencies.  A hid's hid.conf can set a variable "deps"
which is a list of other HIDs that must be included with that hid.  Added ps
the lpr's deps.