Add TO18, TO39, and fix pinout on TO92.
Added TO18 and TO39 footprints based on the JEDEC drawings.  Note that
the pinout used is the one from the JEDEC drawings and it may or may
not match a particular vendors pin numbering.  I have in fact seen
different vendors number the pins differently on these packages.
It is entirely up to the user to make sure that their CAD symbols
have the correct mapping from E, B, C (for a bipolar) to pins 1, 2, 3
and that the JEDEC numbering is used.

While here correct the TO92 pinout.  It now also matches the JEDEC
drawing and is more consistent with the most common vendor numberings.
How on earth has this not caused problems before?