When building from git sources and building docs, force maintainer-mode.
When building from git sources, maintainer mode is required to build
the manual.  So, unless the user has specified --disable-doc then
force maintainer mode.  When building from a tarball, the preformatted
manual is already contained and we do not impose maintainer-mode.
Hopefully this puts to rest the documentation building issues once
and for all!