Adding the ability to use mod1 (alt)(option) in HIDs
This is useful for OSX installs as crtl-click, is mapped to right click.
so when drawing a line that you want to force a DRC "Violation" and you need
to use the control key, it will cancel the line being drawn, and start a
new point.

This is tested in the lesstif and GTK hids, under OSX wher ehte option
key is the "alt" key.   I leave it to the patch integrator th verify on
linux for GTK and lesstif.

This is a bit hackish on detecting the alt key/option key, I #ifdefed it
to __APPLE__ because the option key returns 1<<13 not GDK_MOD1_MASK
Under lesstif I might guess that the alt/mod1 is not quite the same as
option.