Adding ability to customize mouse scroll action in GTK.
Changed the mouse_scroll gtk event handler to use do_mouse_action
instead of manually handling the event.  gtk treats scrolling
differently as a mouse button event, and doesn't give a mouse button
number, so this patch just hard codes up/down/left/right to buttons 4
through 7.  This corresponds to the default names used in the resource
file, and, for X11, gtk hardcodes mouse buttons 4 - 7 to the scrolling
events anyway.  This may cause problems in quartz or windows if some
mouse has a bunch of buttons and quartz/windows maps those buttons to
numbers 4 - 7, but I don't have a system to test that (and in that
case the names in the resource file would be wrong as well).

Added a Scroll action to the gtk hid to be able to mimic the existing
scroll wheel behavior with an action script.