Fix command line exporting.
Previously, hid_actionv was changed to always call gui->get_coords
wether or not the action had a need_coord_msg, in an attempt to always
get the latest coords for things like zoom or pan that didn't have a
need_coord_msg.  However, this broke command line exporting because it
was trying to call get_coords on the nogui hid.

Added the check back in to only call gui->get_coords when the action
has a need_coord_msg and added need_coord_msg's to actions that need
coords.