Allow quoted strings and escaped characters in action arguments.
Quoting works similar to bash quoting:

A backslash (\) is the escape character.  It preserves the literal
value of the next character that follows.  To get a literal '\' use
"\\".

Enclosing characters in single quotes preseves the literal value of
each character within the quotes.  A single quote may not occur
between single quotes, even when preceded by a blackslash.

Enclosing characters in double quotes preserves the literal value of
all characters within the quotes, with the exception of '\' which
maintains its special meaning as an escape character.