Add the start of a testsuite.
Add a testsuite framework.  Currently it is capable of doing tests on
the gerber and bom export HID's.  It should be fairly easy to add png
and nelma export HID's to the testsuite.  The coverage is limited right
now but it should be simple to add additional tests now that the framework
is in place.