win32: Install Readme.txt in @docdir@
Also removes unnecessary fiddling with pkgdatadir which was
responsible for creating an empty directory $datadir/pcb-$VERSION/

Adjusted pcb.nsi.in to match new location of Readme.txt, but have
not test-built the Win32 package due to non-availablility.

Adjusted build_pcb to put @docdir@ at ${pcb_inst}/doc, as this
fits the rest of the win32 build better.