Re-write node_label() to only label the edge being passed as its argument.
Breifly, the old code tried to walk the entire CVCList labeling
multiple edges. The new code tries only to label the argument vertex
i.e. the forward ('N') edge from this vertex. It tries to look
counter-clockwise just one edge which will give the answer in most
cases. If that fails it looks further into the CVCList until it can
find the answer, then stops. The code is much simpler because we know
that the edge we are labeling is always departing the cross-vertex
point.


(Futher comments and original bug diagnosis by Peter Clifton)

This prevents "double labeling", where the old routine could label any
edges which intersect at the vertex passed to node_label().

This fixes a bug discovered by Stefan Salewski, where certain boards
could trigger an assert failure when PCB was built with debugging.


The bug, it would seem - is relating to the way (and order) we were
labeling edges of polygon contours:

2
|/\     In this diagram, x is on the outer contour of our polygon,
x  |    and is in fact two vertices at the same point, lets denote
|\/     these as |\ and |/
1

The labelling algorithm works up from point 1, gets to |\, whereupon it
gets the CVCList for the spatial point x - which contains |\ AND |/
along with the (shared) edges from the clipping contour.

node_label is called with the |\ vertex, but in the process of scanning
the CVCList at this point, it also checks for SHARED edges. It _only_
checks shared edges against the |\ vertex, NOT the |/ one.

During this pass, the | edge coming from the |/ vertex gets mis-labelled
as "OUTSIDE", where is should eventually get labelled "SHARED"

BUG:

We then skip over calling node_label with the |/ vertex, since | has
already been labelled. We next get to point 2.

Point 2 looks at the | edge, and declares it to be "INSIDE". (It should
already have been labelled "SHARED" by this point, so node_label doesn't
know any better.

The | edge is already labelled "OUTSIDE" when we go to label it
"INSIDE".. assertion fail.

This commit avoids missing the correct labelling of shared edges, by
ensuring we never label any edges in advance of calling node_label
for them.