hid/common: Clip no-holes polygon pieces before calling fill_contour
This avoids integer overflow in some HIDs (GTK, Lesstif?) when
drawing at high zoom level. Such overflow would lead to incorrectly
drawn polygons.

It is possible that a similar bug could effect thin-drawn polygons,
but that has not manifested its-self so far. If we were to clip these
in the future, we need to be careful to extend the clip region
slightly off-screen, so the outlines are not drawn.

Ideally we would clip these vertices using a Sutherland-Hodgman
clipping algorithm, then we could simply discard edges which are
clipped completely.