hid/common: Fix memory leak in polygon NoHoles clipping routine
poly_CopyContour creates the edge tree, as does ContourToPoly(),
which overwrite the one created in poly_CopyContour().

Replace call to ContourToPoly() with a discrete alternative without
the poly_PreContour() call.

Also, don't try to poly_Free() the output of poly_Boolean_free() in
an error condition. poly_Boolean_free() re-uses the A input polygon
for its output, so attempting to free that might be dangerous.

poly_Boolean_free() should free its inputs (and any partial output)
if it encounters an error.