By popular(*) request, Make the cross-hair snap to the center of pads
* Popular.. or vocal?, I'm not sure.

If more people don't like the new behaviour, there is always git revert.

There was of course, a third possibility - adding support to snap to pad
end-points _and_ their centers. I rejected this because of "nearly" square
pads - it becomes difficult to know if you've picked up the center of the
pad, and many of those asking for snap-to-center wanted it in order to
pick up and align components by their pads.

NB: Optimised rat-lines still go to the edge of pads, as this can often
    look less cluttered. Rat-lines drawn to the center of a pad will
    work correctly, and remain on the pad center until the next time
    "optimise" is run on the rats-nest.