Rather than having @code{Pcb} all over, refer to the program name with @pcb{}.
To help keep how we refer to and mark up the name of the program, define a
macro at the start and use it all over.

Also change some of the "for X11" text since pcb really runs on what ever
windowing system gtk (X11 or windows) or lesstif (X11 for this one) runs on.