Add support for specifying soldermask and solderside from the command line
Added "solderside" and "mask" as valid layers for --layer-stack.  Also
added "silk" as an alias for "elements".  Added mask support to the PNG
exporter so that it draws soldermask properly.  Also added support to
PNG for showing the board from the solder side, and removed some leftover
cruft from when it was copied from the PS hid.