Remove some gcc warnings about unitialized variables.
Remove some gcc warnings about uninitialized variables.  The
logic was such that if the variables weren't ever assigned to somethign
then they also would not have been accessed but by getting rid
of this warning, we increase the chances of paying attention to
other warnings which may indicate real bugs.