Remove Settings.init_done variable and its usage
This variable was previously used to allow action scripts run prior to
loading the GUI to avoid GUI startup if they executed a "Quit" action.

The requirement dates back prior to the HID split, where the "Quit"
action would cause PCB to call gtk_main_quit(), an operation which is
illegal if the GUI main loop was not yet started. The init_done flag
was used in QuitApplication(), both to determine whether the GUI had
started, and if not - to note that the GUI should not be initialised
later. (Instead of calling gtk_main_quit()).

With the HID split, the QuitApplication() function was changed to call
exit(0) rather than gtk_main_quit(), so this dance isn't necessary at
all now.. in either case, we can just exit the application immediately.