Revert "As silly as it seems to group any sort of scientific or EDA"
I've pushed really hard to sort this nonsense out, and this patch is
not in my opinion suitable for applcation. gEDA, gerbv and PCB all
deliberately omit the Education category.. this is not where we need
to fix this issue.

I strongly object to putting bogus categories into our apps.
If Debian want to patch these in, let them.

This reverts commit b402a144e19225f0648189bca6fbfad95d046047.