Apply patch from Patrick Bernaud:  Provide context to callbacks of actions through global variable.
Providing a context to action is necessary for scripting languages (for
example). To do so, a global variable ('current_action') is being
defined: it is saved, set and restored when running the callback of an
action in a way similar to what PCB already does with HIDs.

The callback function may access the action it belongs to and cast it
the way it wants to extract information from a potentially extended
HID_Action structure.