Don't select silk on the far side if the far side is not shown
The logic for determining if a layer was "on" didn't take into
account that the two silkscreen layers are handled differently.
Thus, if you tried to select shown silk, you would select hidden
silk on the other side too, unknowingly moving that around.

With this patch, silk on the far side is only selected if the
far side ("invisible" side) is shown.