Fix layer visibility logic for some boards.
If the Groups() line listed the c/s layer before the copper layer,
disabling silk would also disable the copper.  This change searches
the entire layer group looking for any layer that might need to be
drawn, and uses that as the exemplar instead of blindly choosing the
first layer in the list.