Fix crash due to uninitalised variable in DRC check Bug #3033772
Andrew Armenia  noted that the second pointer returned from
BuildObjectList(), "object_type_list" would be left uninitialised
if the object type was not encountered. A subsequent call to
free() in DRCFind() would then cause memory corruption.

The issue was originally reported by Ineiev:
http://archives.seul.org/geda/user/Nov-2009/msg00456.html

That patch addresses the lack of a RATLINE_TYPE in the BuildObjectType
function, but does not prevent a segfault for other unknown types.