Fix bug tracker # 3064887
PNG exported all black in photo mode if the "outline" was exactly the
board size, because the left/bottom edges were clipped leaving a big
hole.  With this patch, we do two things:

1. Detect a non-trivial outline, and only do outline processing if
found.  Thus, a rectangular outline exactly the size of the board is
just ignored - no black border on the output image.

2. Any outline line that's drawn exactly on the bottom/right edge is brought in
by half a pixel to ensure it's drawn in the resulting image.