HID-gcode: create better file names.
If the user adds a suffix to a file name in the dialog, he expects
the resulting files to have these suffixes as well.

This patch is a combination of the following two patches by Markus Hitter.

0001-HID-gcode-let-the-system-library-allocate-the-tempor.patch
0004-HID-gcode-create-better-file-names.patch

The original patches are avaiable at the bug-report on
launchpad lp-699497. The combination and modifications where done by Felix
Ruoff.

Affects-bug: lp-699497