gcode HID: implement drill-milling.
This is, if a hole is equal or bigger than the milling tool,
use the milling tool instead of a drill tool to manufacture
the hole. Big advantage is, milling tools can create holes
of arbitrary sizes by moving a circle.

The feature is optional, on by default.

G02/G03 circle movements are avoided, as not all machine
controllers understand them. Instead, a polygon is milled,
with enough sides to stay for any hole size  within the user
choosen accuracy.

If the hole is only slightly bigger than the mill tool, don't
move the circle; a simple plunge is enough.