SF patch 1674895:
This crash occurs in rats.c, around line 610 in the function
DrawShortestRats() where CreateNewRat() is called with some arguments
which don't exist.

It occurred for me when I tried to "optimize rats nest" on an invalid
design.

There's no check in DrawShortestRats() that the firstposition and
secondposition pointers are not NULL.
firstposition->X (and Y) and secondposition->X (and Y) are passed as
arguments to CreateNewRat(). It is possible in some cases that
firstposition and secondposition are not reset from 0 in the code
preceding the call to CreateNewRat(). In these cases a segfault will
occur.