Revert undo changes relating to RatFindHook calls
This reverts commit 1aac06b665330507ca42f380e38f32875a979cd9
         and commit d3b508ca272b7d89aa27f552ab78943f9bed27f9.

I made a mistake by assuming the undo data saved by the RatFindHook
(when looking up nets to select) was redundant. I had presumed this
since the flags being set are reset again before we finish processing.

What actually happens, is that further processing sets the SELECTED flag
on the FOUND objects. Undo data is saved for the SELECTED flag addition,
which stores the previous flags of the object. When an undo is performed,
it also restores the FOUND flag we wanted to keep unset.

In order to get back to our original state, we need the RatFindHook to
store undo information for when the FOUND flag was initially set.

The ResetConnections() call after the RatFindHook should still be ok
not saving undo information, as it only touches object flags which
will be restored anyway by the undo data now saved by the RatFindHook.

NB: Revert was manually modified to retain an indenting change in find.c
and to add the required changes to select.c, which had additions using
the RatFindHook API since the above reverted commits were made.