Allow zoom-out-past-board for Lesstif
The lesstif GUI has been modified two ways:

1. You may zoom out "past" the board, making the board smaller
than the window.  Note that the crosshair is still limited to
the board's extents.

2. You may pan the board so that edges need not be on the edge
of the window, such as centering an edge for inspection.

Note that the smallest the board can be is 1/10th the window
size.  While this is arbitrary, it's easily changed if needed,
and avoids problems with miniscule boards (and thus maxiscule
coordinate systems).

The 'v' key still zooms the board to fit, but now centers the
board in the window rather than placing it at the upper left.