Fix "select pads on hidden far side"
If an SMT pad is on the far side of the board, and the far side
layer is not visible, do not select that pad.  Exception: if that
pad is part of an element which *is* on the visible side, the pad
is selected as part of selecting the whole element.  But, elements
on the far side (the common case) won't have their pads selected
when those pads aren't visible.