(GTK HID) fix annoying decimals
When in metric grid mode, the coordinates are displayed
with irregular tails of nines, e.g. "116.9998 43.9999"
when the grid is set to 1mm.

The fix includes using more precise metric-to-imperial
and crosshair calculations and fitting crosshair values
into metric grids.

To address the issue in a really reliable way we need at least
one more separate variable, to explicitly mark the grid
as metric (as opposed to the displayed units);
this flag should also be saved in PCB files.

First reported by Kai-Martin Knaak.

Closes-bug: lp-699452 (sf-2117383)
Closes-bug: lp-699169 (sf-1741659)