hid/gtk: Select adequate folder in file-save-dialog
Since now, the 'Save layout as' - dialog gives full path and filename in
name-line, but did not switches to given path in folder-list (lower part of
the dialog). For me, this is uncomfortable, e.g. if I just want to switch to
another near-by - folder.

This solution was suggested by Krzysztof Kościuszkiewicz. Thank you!

Closes-lp: 699491
Reviewed-by: Krzysztof Kościuszkiewicz <k.kosciuszkiewicz@gmail.com>