hid/gtk: expand library tree after filtering
This patch changes the behavior of the library selection dialog.
After filtering by name is applied the tree view of matching components
is automatically expanded so that all leafs are immediately selectable.
The tree view is collapsed when the filter is cleared.

Patch is adapted from Krzysztof Kosciuszkiewicz's patch for gschem.

Closes-bug: lp-753643
Reviewed-by: Krzysztof Kościuszkiewicz <k.kosciuszkiewicz@gmail.com>