Fix warning for gtk describing menuitems
With GTK+ there are no 'describing menuitems' avaiable (the out-grayed items
in lesstif-GUI). For this reason, the items in gpcb-menu.res are ignored and
a warning is displayed at the message-log.

This patch fixes this by removing these describing menuitems for the gtk-GUI
and renamed the described items to a more explaining label.
It also added mnemonics for these menu-items.

Closes-bug: lp-769815