puller.c: Re-write parts to use hash tables to store stashed extra info
This vastly simplifies some of the pointer arithmetic which was previously
used to account for underlying location changes when objects moved.

Having tested this, it produces different results than before application.
This may be due to it zapping some underlying bug which has been removed
in the patch, or perhaps more likely, that its slightly different iteration
order over the lines and arcs (due to g_hash_table_foreach) will result
in different pathologies being hit.