Add raw polygons support for the GTK+GL HID, bypassing the no-holes dicer
Uses the OpenGL stencil buffer to make drawing polygons with holes faster.

1. Turn Stenciling on, updates to colour buffer off
2. Clear stencil buffer to 0
3. Paint polygon holes, setting those areas of the stencil buffer to 1
4. Switch on stencil test (== 0), turn on updates to colour buffer
5. Paint outer polygon through areas of the stencil buffer still 0
6. Clear stencil buffer, switch off stencilling.

Caveat:

This function might throw up if it is used whilst drawing the mask, since
that uses stenciling as well. We don't use polygons on the mask, so its
not a a problem. (Mask cutouts for octagonal pins do work correctly).