Lesstif + GTK HIDs: Ignore ISO Level3 shift modifier key
This key press is found on some international keyboards (e.g. Spanish).
If we don't ignore it, PCB complains about the key not being tied to
an action when it is pressed.

From the committer, Peter Clifton <pcjc2@cam.ac.uk>
...
Thanks to Gabriel for reporting this and providing the patch for
Lesstif. I have updated the patch to make the equivalent change to
the GTK HID as well.

Signed-off-By: Peter Clifton <pcjc2@cam.ac.uk>