Don't register the PNG hid if there are no file formats to use.
In some cases, the GD library may not have any usable image formats for
us to use, resulting in an empty filetypes[] table.  Rather than trying
to accomodate an empty table during export, it's easier (and perhaps
better) to simply not register the PNG hid in that case.  Thus, the
exporter is simply not available when the table is empty.