Fix shift/reduce conflicts
Closes-bug: lp-794743

I moved the interpretation of what "measurement" means even further up
the heirarchy, by storing the original integer part as well as the units
in a structure, and applying the new/old/unitless meanings where they're
used.  This way, there's no conflicts between a number which is a
measurement, and a number which is a flag or mask, for the old formats
that differ only in the number of parameters.

I also cleaned up the parser rules for "zero or more" type lists,
to remove most of the remaining conflicts.  There's still one at the
toplevel because we try to parse different types of files with the same
parser.

Lastly, the symbol rules were combined into a "symbol head" rule with a
common "symbol body" rule.

As a side effect, we can now reliably use units on old-style formats
again, since the *100 scale factor is only applied if the units are not
specified.