Set file pointer to NULL after closing in gerber.c
Closes-bug: lp-795734

When cleaning up the gerber HID, I had changed the
file pointer used in maybe_close_f from a global to
a (local) argument -- which made the line ``f = NULL''
effectively a no-op.

However, the global variable f does need to be set
to NULL for the code to recognize that the file is
closed; otherwise it tries to keep using the (now
invalid) file handle, causing crashes whenever the
gerber HID is used more than once.

This is now done after every call to maybe_close_f(),
rather than depending on the function itself to have
side effects.