Add %m+ specifier to pcb-printf
As long as our base units are cmils, allowing pcb-printf to guess
the most natural units for displaying measures is a bad idea -- it
cannot reliably count significant figures for many values. The
result, for example, in report.c, is inconsistent and confusing
dialogs with a mix of metric and imperial measurements.

The %m+ specifier is used to force pcb-printf to only use certain
units; in this case, the user's gui setting of metric/imperial.