Change report.c to use pcb-printf
Right now there should be no change in the output from
report.c; this is essentially just a code cleanup. The
prec/UNIT nastiness is cleaned up -- precision is handled
by the defaults in pcb-printf (.2 for mil, .4 for mm,
same as before) and unit selection is done with the %m+
specifier.

It's easy now to expand to allow auto-scaling or even
automatic selection of metric/imperial based on sig.
figs., but I've played around with this and it looks
like mixing units is confusing and hard-to-read, so we're
sticking with just one of mm/mil for now.