Convert gtk gui to use pcb-printf
There should be no user-visible changes.

Note that as long as base units are cmils, the .9999 bug
in the cursor status is back. This is not a problem with
smaller base units (I am using 17nm and all is okay.)

Note also that the report dialogs are entirely based on
report.c, so that is not part of this commit.