Make file.c use %mr pcb-printf spec
I have changed the %mr spec to always output cmils,
no suffix, and changed file.c to use this. The reason
is that the %mc spec (cmils, no suffix) is locale-
dependent, while %mr is not.

When we change the actual file format, file.c can be
left alone and the relevant changes should be done to
the %mr spec in pcb-printf.