Remove mm/mil dichotomy, support arbitrary user units
Currently, pcb assumes the user's display units are either mm
or mil, and uses the Settings.grid_unit_mm boolean (and flag
"grid_unit_mm") to determine which is which.

This patch removes the boolean and replaces it with the new
Settings.grid_unit, which can be set to any unit supported by
pcb-printf. The user-visible interface has not changed (there
is still a mm/mil toggle in Gtk and the menus only contain mm
and mil units), but new units can be accessed though the
:SetUnits command.

The flag is still there and can be used in pcb-menu.res as
usual. However, the new flag "grid_unit_mil" should be checked
to see if mils are selected, since this is no longer implied
by grid_unit_mm == 0.

There will be some user-visible changes to the precision of
displayed values, since I have removed a lot of special-case
code for this sort of this thing and use the default_prec
of applicable units instead.

Because of the new flexibility, some idioms have been changed:

  Settings.grid_unit_mm ? COORD_TO_MM (x) : COORD_TO_MIL (x)
becomes
  coord_to_unit (Settings.grid_unit, x)

  Settings.grid_unit_mm ? "mm" : "mil"
becomes
  Settings.grid_unit->suffix

  Settings.grid_unit_mm = 1;
becomes
  Settings.grid_unit = get_unit_struct ("mm");

For GUI use, the Unit structure returned by get_unit_struct
exposes certain members:

  suffix       : "mm"/"mil"/etc
  in_suffix    : i18n version of the above
  default_prec : precision used for spinboxes, labels, etc

  step_tiny    :
  step_small   :
  step_medium  : step sizes for various spinboxes
  step_large   :
  step_huge    :

Additionally, the *_increment_mm and *_increment_mil variables
have their own structure containing default, min and max values.
These can no longer be set on the command line.