crosshair.c: Drop checks to ensure the crosshair is inside the PCB
It would seem that the MIN, MAX functions applied when setting
Crosshair.X and Crosshair.Y should already constrain the coordinates
such that the removed code was never triggered.

It isn't obvious that these removed checks are even correct, or are
being triggered. Drop the checks in the hope that the now simplified
code continues to "just work" without the special casing.