line.c: Fix a bogus compiler warning regarding a (X + c) >= X being always true.
line.c: In function ‘EnforceLineDRC’:
line.c:491:7: warning: assuming signed overflow does not occur when assuming that (X + c) >= X is always true [-Wstrict-overflow]

It would appear that gcc (4.6.1-5ubuntu1) is identifying the conditional
portions of INDEXOFCURRENT which would always evaluate a particular way
if a given condition is met. It is "probably" safe to assume that this
warning manifests due to gcc considering each portion of the INDEXOFCURRENT
definition in turn.

It appears we can avoid this particular warning by temporarily assigning
INDEXOFCURRENT into a variable before we test with it.