Introduce PCB::grid::unit attribute
This is PCB's first use of the Attribute() field in the
file format. It is a unit suffix string denoting the unit
setting used by pcb when loading the file.

Note the namespacing: as Attributes are persistent across
file-saves, other programs may use them in future for
purposes unknown and irrelevant to pcb. Therefore we will
put all pcb attributes under the PCB namespace.

If this attribute is missing or invalid (i.e., the unit
given is unsupported by pcb-printf), PCB will then use the
--grid-units command-line option. Failing that, it will
use the grid-units entry in ~/.pcb/preferences. Failing
that, it will use mils.

Fixes-bug: lp-811393