Correct parameter order in Distance function in hid/lesstif/main.c
in src/did/lesstif/main.c, the Distance function is called
with the parameters in the wrong order.

This small patches fixes the problem. I still think that
the format is wrong, or at least the rounding to integer
mm/mil taking only the grid into account, but I've not yet
decided on how to fix it. The worse is that often the display
is wrong because the decimal part is truncated (displaying
say -24 between two grid points on a 5mil grid).

        Gabriel