Modify get_unit_struct()'s bad-unit forgiveness
Suffixes passed to get_unit_struct() may now start or end
with whitespace, without affecting the result. However,
incomplete units will NOT be matched.

This means that "mi" will no longer return the "mil" struct,
for example.

The reasons for this change are:
  1. The old behavior returned the first potential match,
     regardless of other matches: "c" is always "cm", never
     "cmil".
  2. Prevent surprises (due to point #1, or typos).
  3. Prevent user dependence on behavior that will change
     as units are added or removed.

It still supports plural units, like "inches" or "mils".
However, it will read "miles" as "mil" because of this. ;)