Introduce GtkPcbCoordEntry widget, use it in gui-config.c
The GtkPcbCoordEntry is a modified spinbox that handles
pcb units internally and outputs them as human units. It
uses the step sizes given in pcb-printf and adjusts
automatically when units are changed.

If you manually change the unit suffix, it will change
its interal unit, so that if you change "10mil" to "10mm"
it will do the right thing.

TODO: handle overflows
      add a unit selector to the context menu