refdes labels in new layout can't be moved.
The .pcb file that gsch2pcb creates does not have a font in it.  When
loading files, the bounding box for all the text is calculated as the
file is read.  In the case that there is no font, this is before the
default font is installed, so the bounding box is way too small.

This patch remedies this by making all new PCB structures contain the
default font by calling CreateDefaultFont in CreateNewPCB
(CreateDefaultFont now takes in a PCBTypePtr instead of using the
global PCB).  Previously, each time CreateNewPCB was called,
CreateDefaultFont was called soon thereafter.

In the case of loading a PCB from a file.  The default font in the
newly created struct PCB is marked invalid but not removed.  If the
.pcb file contains a font, the default font will be overwritten by the
file's font.  If it does not, PCB->Font will still be invalid after
LoadPCB, a message is displayed that the default font is being used,
and PCB->Font.Valid is set to true.

Also fixes a related bug where the memory for a symbols lines wasn't
being cleared when a new file was brought into place.  Previously,
this only happened if a new Font was loaded into an existing PCB.  Now
the font will be replaced every time a file with symbol information is
loaded.

Rebased to actual git-head by Felix Ruoff

Closes-bug: lp-699478